requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
openpyxl>=3.0.0
tabulate>=0.9.0
flask>=3.0.0
flask-cors>=6.0.0

